# Refactor Dockerfile