This PR refactors the dockerfile according to issue https://github.com/ElementsProject/lightning/issues/8409

changes:
- [x] remove useless python-builder stage
- [x] verify GPG keys (https://github.com/ElementsProject/lightning/issues/8405)
- [x] only include the necessary binaries in final image
- [x] use of ARG instead of ENV for build-time vars
- [x] remove useless apt packages
- [x] use of `ADD` instead of `curl`
- [x] set global `SHELL`
- [x] reduce redundancy of multi-arch image selections
- [x] use of `tee` + heredocs instead of messy echo's
- [x] stripped compiled binaries for smaller image
- [x] remove manual `tini` in favor of docker `--init` flag
- [x] avoid compiling sqlite, zlib, postgres from source
- [x] replace poetry with uv
- [x] avoid downloading all history/branches for submodules
- [x] parallelize `make` and submodules cloning 
- [x] better docker caching of layers
- [x] fix CPU compatibility bug (https://github.com/ElementsProject/lightning/issues/8456)
- [x] avoid running cargo with QEMU. true cross compilation.
- [x] avoid installing `clippy` and `rust-doc` components 
- [x] any host to any target image building (not tested)
- [x] avoid installing manpages
- [x] update base image from bookworm to trixie
- [x] fix poetry trying to install the current project as standalone package (crashed the build before)
- [x] fix armhf target (crashed the build before)


| Feature        | Before       | After        | Improvement |
|----------------|-------------|-------------|---------------|
| Build Time  (amd64 to arm64)   | 90.6 min      |   10.2 min    | 8.8x          |
| Image Size (amd64)    | 512 MB       | 247 MB       | 2.07x           |
| Lines of Code  | 316         | 199         | 1.6x           |

build time was benchmarked on a 12th Gen Intel(R) Core(TM) i5-1235U and a 20 Mbps connection