@ShahanaFarooqui 2025-07-31
@Raimo33 As per @cdecker's suggestion [here](https://t.me/lightningd/37125), we should modify the Dockerfile to run as a non-root user for improved security. Since we're already refactoring the Dockerfile, could we incorporate this change as well?