@Raimo33 2025-08-02
I'm removing the manual handling of `tini` in favor of a user controlled approach, more docker native and friendly.

```Dockerfile
ENTRYPOINT  [ "/usr/bin/tini", "-g", "--", "./entrypoint.sh" ]
```

becomes:

```Dockerfile
ENTRYPOINT  ["/entrypoint.sh"]
```
and the user is free to run the container with tini by using the official `--init` flag as follows:
```bash
docker run --init ...
```