@Raimo33 2025-08-06
The Dockerfile now should also support building the image from any **host** architecture as well. It previously only suppported amd64 to any. now it should be any to any, but I've no way of testing it