@Raimo33 2025-08-11
also I wonder why our Makefile `make install` checks the availability of avx intrinsics:

```bash
#80 511.3 checking whether C compiler accepts -msse3... no
#80 512.0 checking for SSE3 instructions set... yes
#80 512.0 checking whether C compiler accepts -Wformat-nonliteral... no
#80 513.1 checking whether C compiler accepts -mssse3... yes
#80 513.6 checking whether C compiler accepts -Wformat-security... no
#80 513.9 checking for SSSE3 instructions set... no
#80 514.9 checking whether C compiler accepts -msse4.1... yes
#80 515.1 checking whether C compiler accepts -Wformat... no
#80 515.6 checking for SSE4.1 instructions set... yes
#80 516.3 checking whether C compiler accepts -Wstrict-prototypes... no
#80 517.0 checking whether C compiler accepts -mavx... no
#80 517.7 checking for AVX instructions set... yes
#80 517.8 checking whether C compiler accepts -Wshadow... no
#80 518.8 checking whether C compiler accepts -mavx2... yes
#80 519.1 checking whether C compiler accepts -Wnested-externs... no
#80 519.6 checking for AVX2 instructions set... yes
#80 520.3 checking whether C compiler accepts -Wcast-align... no
#80 520.5 checking whether C compiler accepts -mavx512f... no
```

For a generic -- even arch specific -- docker build, we shouldn't use any intrinsics. They are not guaranteed to be present on all CPUs. There are different kinds of arm64 CPUs and different kinds of amd64 CPUs.