@Raimo33 2025-08-12
An implementation of https://github.com/ElementsProject/lightning/issues/8443 would speedup the build quite a bit