@Raimo33 2025-08-21
When compiling for armv7 I'm getting this unaligned access error:

```bash
#49 7.927 checking for unaligned access to int... ccan/tools/configurator/configurator: Test for HAVE_UNALIGNED_ACCESS did not compile:
#49 7.959 configuratortest.c: In function 'main':
#49 7.959 configuratortest.c:8:22: error: array subscript 'int[0]' is partly outside array bounds of 'char[4]' [-Werror=array-bounds]
#49 7.959     8 |         return *x == *y;
#49 7.959       |                      ^~
#49 7.959 configuratortest.c:5:11: note: at offset 1 into object 'pad' of size 4
#49 7.959     5 |      char pad[sizeof(int *) * 1];
#49 7.959       |           ^~~
#49 7.959 cc1: all warnings being treated as errors
#49 7.960 
#49 ERROR: process "/bin/bash -euo pipefail -c ./configure --prefix=/tmp/lightning_install --enable-static --disable-compat --disable-valgrind" did not complete successfully: exit code: 3
```

This is because armv7 does not guarantee unaligned access support, so the compilation should rightfully fail. Now i wonder, are we sure that this repo is meant for armv7 as well?