@Raimo33 2025-08-21
@cdecker Adding a non-root `USER` directly inside the Dockerfile is not a reliable option when using mounted volumes. The reason is that file permissions on mounted volumes are determined by the **UID and GID of the host system**, not the container. Hardcoding a UID, for example `USER 1000:1000`, might work in some environments, but it is **not guaranteed** to match the host system’s user. On many systems, the first regular user is UID 1000, but this is not universal, and in environments like CI/CD runners, containers, or systems with different user configurations, the UID/GID may differ.

To safely allow a container to access a mounted volume without running as root, the UID and GID of the container user must either:

1. Be dynamically adjusted at runtime to match the host’s UID/GID, or
2. Be explicitly specified by the caller via `docker run --user $(id -u):$(id -g)`.

Hardcoding the UID inside the Dockerfile creates a fragile assumption and can easily lead to permission issues. Some repos such as [mempool](https://github.com/mempool/mempool/blob/master/docker/backend/Dockerfile) do just that, but I don't deem it robust.

Therefore I propose leaving it as is, and letting the runner speficy `---user`. Same solution as the `--init` flag for tini.