@Raimo33 2025-08-26
Closes https://github.com/ElementsProject/lightning/issues/8409 https://github.com/ElementsProject/lightning/issues/8405 https://github.com/ElementsProject/lightning/issues/8456 https://github.com/ElementsProject/lightning/issues/5655