@ShahanaFarooqui 2025-09-18
> Agreed that the dependence on `git describe` is suboptimal, do you have an alternative we should use? Maybe `git describe` if there is no file .METADATA on disk? Otherwise `source .METADATA` which should define a VERSION var?

Can we not utilise `./.version` file which is updated with `make update-versions` every time before tagging the release?