@cdecker 2025-09-22
We definitely could, however I think that would be worthwhile to add in a followup. This PR is big (and helpful) on its own, and I should not have brought up the possibility without mentioning that it is very much nitpicking, and not required for this PR to proceed. Sorry @Raimo33 for keeping this pending as long as we have, this is excellent and we should have merged it already. I'll do one last round, and merge it asap 🤗