@Raimo33 2025-09-22
Thank you for the kind words, I haven't tested thoroughly the plugins with this refactor as I'm not too familiar with the codebase and lightning from a technical point of view, if you haven't done so, please test before ACKing.